テストを集約
